requirements.txt
gradio
pymupdf
requests
python-dotenv
markitdown[pdf]
pillow